Introduce dynamic page => more control on page hierarchy, fix bugs

+++ Command/CacheFlushAllCommand.php
@@ -24,13 +24,21 @@ class CacheFlushAllCommand extends Command
     {
         $this->setName('sonata:page:cache-flush-all');
         $this->setDescription('Flush all information set in cache managers');
+
+        $this->addArgument('manager', InputArgument::REQUIRED, 'The CMS manager to use');
     }
 
     public function execute(InputInterface $input, OutputInterface $output)
     {
         $output->writeln('<info>clearing cache information</info>');
 
-        foreach ($this->getManager()->getCacheServices() as $name => $cache) {
+        $managerName = $input->getArgument('manager');
+
+        if (!in_array($managerName, array('snapshot', 'page'))) {
+            throw new \RunTimeException(sprintf('Please provide a valid provider : snapshot or page'));
+        }
+
+        foreach ($this->getManager($managerName)->getCacheServices() as $name => $cache) {
             $output->write(sprintf(' > %s : starting .... ', $name));
             if ($cache->flushAll() === true) {
                 $output->writeln("<info>OK</info>");
@@ -43,10 +51,11 @@ class CacheFlushAllCommand extends Command
     }
 
     /**
-     * @return \Sonata\PageBundle\Page\Manager
+     * @param string $manager
+     * @return \Sonata\PageBundle\CmsManager\CmsManagerInterface
      */
-    public function getManager()
+    public function getManager($manager)
     {
-        return $this->container->get('sonata.page.cms.page');
+        return $this->container->get('sonata.page.cms.'.$manager);
     }
 }
